add http status and handle exceptions (post and category)

+++ src/main/java/com/dauphine/blogger/services/impl/PostServiceImpl.java
@@ -1,7 +1,12 @@
 package com.dauphine.blogger.services.impl;
 
+import com.dauphine.blogger.dto.CreationPostRequest;
+import com.dauphine.blogger.exceptions.CategoryNameAlreadyExists;
+import com.dauphine.blogger.exceptions.CategoryNotFoundByIdException;
+import com.dauphine.blogger.exceptions.PostNotFoundByIdException;
 import com.dauphine.blogger.models.Category;
 import com.dauphine.blogger.models.Post;
+import com.dauphine.blogger.repositories.CategoryRepository;
 import com.dauphine.blogger.repositories.PostRepository;
 import com.dauphine.blogger.services.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,21 +19,17 @@ import java.util.UUID;
 public class PostServiceImpl implements PostService {
 
     private final PostRepository repository;
+    private final CategoryRepository categoryRepository;
 
-    public PostServiceImpl(PostRepository repository) {
+    public PostServiceImpl(PostRepository repository, CategoryRepository categoryRepository) {
         this.repository = repository;
+        this.categoryRepository = categoryRepository;
     }
 
     @Override
-    public List<Post> getAllByCategoryId(UUID category_id){
-        List<Post> allPosts = this.repository.findAll();
-        List<Post> posts = new ArrayList<>();
-        for (Post post : allPosts) {
-            if(post.getCategory_id().getId()==category_id){
-                posts.add(post);
-            }
-        }
-        return posts;
+    public List<Post> getAllByCategoryId(UUID category_id) throws CategoryNotFoundByIdException{
+        Category category = categoryRepository.findById(category_id).orElseThrow(() -> new CategoryNotFoundByIdException(category_id));
+        return repository.findAllByCategoryId(category_id);
 
     }
     @Override
@@ -42,27 +43,32 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public Post getById(UUID id){
-        return repository.findById(id).orElse(null);
+    public Post getById(UUID id) throws PostNotFoundByIdException{
+        return repository.findById(id).orElseThrow(() -> new PostNotFoundByIdException(id));
     }
     @Override
-    public Post create(String title, String content, UUID category) {
+    public Post create(String title, String content, UUID category_id) throws CategoryNotFoundByIdException {
+        Category category = categoryRepository.findById(category_id).orElseThrow(() -> new CategoryNotFoundByIdException(category_id));
         Post post = new Post(title, content, category);
         return repository.save(post);
     }
     @Override
-    public Post update(UUID id, String title, String content) {
+    public Post update(UUID id, String title, String content) throws PostNotFoundByIdException{
         Post post = getById(id);
-        if (post == null) {
-            return null;
-        }
+
         post.setTitle(title);
         post.setContent(content);
         return repository.save(post);
     }
     @Override
-    public boolean deleteById(UUID id) {
-        repository.deleteById(id);
-        return true;
+    public boolean deleteById(UUID id) throws PostNotFoundByIdException {
+        boolean exists = repository.existsById(id);
+        if (exists) {
+            repository.deleteById(id);
+            return true;
+        } else {
+            throw new PostNotFoundByIdException(id);
+        }
+
     }
 }
